build_map/src/requirements.txt
geopandas
shapely
pyproj
rtree
numpy
pandas
matplotlib
scikit-learn
pillow
